If exposure time 0 is set, then use current shutterspeed value
For example, the current shutterspeed value is set to 5 seconds
and the exposure time is set to 0 seconds. The camera exposures
for 5 seconds although an exposure time of 0 seconds was specified.
In other words, triggering an exposure of 0 seconds actually,
triggers an exposure of current shutterspeed value.

In addition, show the libgphoto2 version when driver is started.