Changed ccd, guider, and wheel asi drivers to arm6
The static library files for ccd_asi, guider_asi, and wheel_asi
(libASICamera2.a, libUSB2ST4Conv.a, and libEFWFilter.a respectively)
were generated for arm7 architecture. This made the compiled drivers
unusable on arm6 devices (i.e. Raspberry Pi 0/1). Since the arm7
architecture is mostly backwards compatible, the static libraries were
changed to their arm6 version. The libASICamera2.a and libEFWFilter.a
files were obtained from the ZWO SDK, and the libUSB2ST4Conv.a was
obtained from the libindi repository.

Resolves: #164