Support new protocol response for port state
The protocol for the UPBv3 "IS" command has changed.

Old format:
IS:0:0:0:0:0:0
with each value representing port state.

New format:
IS:0:0:0:0:0:0:000000
The first values separated by ":" represent the port current, in mA.
The final string of 6 digits represents the port state.

Seen on firmware v 1.4

As part of the fix, I added properties for the current.
Even when the ports are off, the output current reading fluctuates with a few milliamps, therefore I decided to round (down) the nearest 10 mA.